Renaming App container to Home container
Home container will be the home screen the user first sees when the app is opened. This will do the fetching of the conversion rates and
update the redux store.
This update includes the addition of more tests for the action creators as well as updating how the reducer will update the redux store.